+++ frontend/src/hooks/useZeroRound.tsx
@@ -1,12 +1,23 @@
 import { useEffect, useState } from 'react';
 import { emitSubmitQuizReply } from '@game/NetworkServiceUtils';
 import { useRecoilValue } from 'recoil';
-import { quizReplyState } from '@atoms/game';
+import { quizReplyState, roundNumberState } from '@atoms/game';
+import styled from 'styled-components';
+import { Center } from '@styles/styled';
 
-function useZeroRound(curRound: number) {
+function useZeroRound() {
     const placeholderDefault = '그림을 보고 답을 맞춰보세요!';
     const quizReplyContent = useRecoilValue(quizReplyState);
+    const { curRound } = useRecoilValue(roundNumberState);
     const [placeholder, setPlaceholder] = useState(placeholderDefault);
+
+    const renderZeroRoundGuide = () => (
+        <ZeroRoundGuide>
+            나만의 문장을 만들어 입력해보세요!
+            <br />
+            다른 사람들이 어떤 그림을 그리게 될까요?
+        </ZeroRoundGuide>
+    );
 
     useEffect(() => {
         setRandomWordToPlaceholder();
@@ -29,7 +40,15 @@ function useZeroRound(curRound: number) {
         emitSubmitQuizReply({ quizReply: { type: 'ANSWER', content: quizReplyContent } });
     }
 
-    return { placeholder, sendRandomWordReplyToServer };
+    return { placeholder, sendRandomWordReplyToServer, renderZeroRoundGuide };
 }
 
 export default useZeroRound;
+
+const ZeroRoundGuide = styled(Center)`
+    ${({ theme }) => theme.layout.sketchBook};
+    height: 420px;
+    color: ${({ theme }) => theme.color.primaryLight};
+    font-size: ${({ theme }) => theme.typo.h3};
+    font-weight: 600;
+`;
